backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.135.1",
    "motor>=3.7.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
